Cleanup last import cycle from ldap connector to userdb
The load_users/save_users functions are not moved to the utils module
at the moment, because we would have to drag too many helper code with
them. As temporary(tm) workaround we hand over the function pointers to
the do_sync method.

Would be good to move the whole stuff to some kind of object that can
be handed over to the connectors using the constructor. Would be much
cleaner.

Change-Id: Icf008d196db92bd9226f2ae749ae3fb483c726f7