Group data source tests per module
Before refactoring the data sources, we should have more tests.  This
new organization should make it easier to test and modify the data
sources independently.

This patch only moves code around.  The only other change is the removal
of the namespaces in order to better identify what is currently tested.

CMK-3742

Change-Id: Ic6291bfe1a4d27729589cd6516eb5d5bebc59347