Split TCPDataSource::_execute()
The method is long and has many responsibilities: gathering the
configuration, fetching and decrypting the raw data, and error handling.
This patch makes these steps explicit: the code is now more procedural
(this is a good thing: Dijkstra would be proud).

We further make it clear that the data fetching part is independant from
the class by making the extra methods static.

CMK-3742

Change-Id: I3577c53cc5ec8804a5cab96219ef0ceb36a16225