Remote checkmk automation calls can now be executed non blocking
These remote automation calls are used by WATO for inter site
communication. Each time the central WATO wants to execute an action on
a remote site, these automations are used.

To mitigate the problem, these calls can now be made non-blocking. In
this case, processing is started with an HTTP call and waits for the
process to be completed with subsequent HTTP calls. As soon as a result
is available, it is reported back to the caller.

The feature can be used for all Checkmk automations, but is initially
only used for bulk discovery calls, which is activated in the following
commit.

Change-Id: I3f0f3bc9274e9ad5654b7cbd8c33c7d245492ade