10875 FIX Services: Escape service details equal to status views
The service details (aka plugin output) was escaped in an unusual way
on the service discovery page. It allowed some tags to be rendered
without escaping (internally known as permissive encoding).

Even if it was not really problematic, at least it confused users. So now the
escaping is done the same way as in the status views.

Change-Id: I59db0e9af85094b4b2962421bff2a16fa7f7e575