10777 New metric for percentage of used filesystem
The filesystem check mentions the filesystem usage as a percentage but did
not store that metric. With the inclusion of "Metric History" painters,
accessing this percentage value becomes increasingly relevant when trying
to compare hosts.

To keep access times lows, this metrics is now generated by the filesystem
check.

CMK-4148

Change-Id: Ifc4590a4f10a6ee1a54ccdbc28c2304761b2cb65