Make IPMIDataFetch a class again
Create a few useful fields and remove the "staticmethod" decorators.

Also, the tighter focus of the class allows us to shorten the name of
the methods.  For example, a class called IPMI*Fetcher does not need to
prefix its methods with `fetch_ipmi`, etc.

We are getting somewhere.  The only thing missing is a context manager
that allows us to simplifies and disentangle IPMIDataFetcher.data()

Also: Do not build strings before passing them to the logger.

CMK-4143

Change-Id: Ia6e4b69a5b557a8e8b4ae4907715f982bf4a9365