Don't load scripts in snapins multiple times
The switch to iframes introduced a bug that loaded
script tags in sidebar snapins multiple times. If
a sidebar snapin like the Service Speed-O-Meter
registers a timer, it was registered multiple
times if other snapins are added and the to the
sidebar and the Service Speed-O-Meter was already
present in the sidebar.

Change-Id: Id74322bf45895744028593b55b490ce33f5a1aca