GUI: Fix forward slash escaping
This escaping is optional according to the JSON spec and not done by the
json.dumps() of Python 2. Since we add some of our JSON objects to HTML
code all over the place it makes sense to apply this by default in the
GUI code.

We should really find a cleaner implementation, e.g by creating a custom
JSONEncoder class.

Change-Id: If3ac72d38597b5e4aaaea22eec7af49fa5c6121d