Do not delay construction of summary
The set is filled in the fetcher and used in the helper.  This would
imply passing it back from the fetcher and, therefore, some sort of
serialization.  This can be avoided by getting rid of the intermediate
structure.  Indeed, delaying the construction of the ServiceCheckResult
is not necessary: the intermediate values kept in the set are never used
by themselves.

CMK-4143

Change-Id: I795baf0654d4a7d8c89d0878aacbfe6ee0530ea7