10125 FIX Fixed separator of SNMP trap messages
The event console messages created from SNMP traps can contain a list of
OID/value pairs.  These pairs were separated by " , " previously, now they
are separated by ", " (not the missing space before the comma).  Although
the new behavior is more correct, it can have an impact on your configured
event console rules: If you match on the text of the messsage, the change
can make a difference, depending on the regex you use.

A portable fix is to use " ?, " to match the separator, i.e.  make the first
space optional in the regex.

Change-Id: I7a47c50f16b76edbd5487809af184e6537556f8e