10882 FIX Fix "Customers" dashlet shown on customer sites
The customer sites only have information for a single customer.
It does not make sense to show the customers snapin in the main
dashboard. You can view the numbers and open the related objects
using the Tactical overview snapin.

The customer snapin is now removed automatically from the builtin
main dashboard. In case you have built a custom dashboard, it will
not be affected by this change.

Change-Id: Iadd7b7393925cf704ae2a44fd007e50c752d2582