apc_symmetra.temp: prepare to 'inherit' items from apc_symmetra_temp
With the new check API we won't have subchecks anymore,
and the '.' in the names will be replaced by '_'.

This commit deals with the only resulting conflict of
the subcheck 'apc_symmetra.temp' versus the main check
'apc_symmetra_temp'.

It prepares the subcheck to "take over" the services
of the main check, once we start interpreting all
dots in names as underscores. If that moment has come,
we will remove 'apc_symmetra_temp', and start discovering
all other sensors in the now-subcheck.

Change-Id: I0aa2e8966653a4f828c12868137f7160e9364245