11105 apache_status: Configuration of Instance Names
It is now possible to configure names for instances in three ways:

 - Autodetection according to OMD configuration
 - Overwrite of automatically detected instances
 - Setting an Instance Name for manual configured

If ENABLE_OMD_SITE_DETECTION is set to 'True' in apache_status.cfg, the names
for server instances are translated to OMD site names if the ip and port match
the omd configuration.

To overwrite instance names by preconfigured values, use CUSTOM_ADDRESS_OVERWRITE
in apache_status.cfg. Use hostname/IP:Port as key, the new name as the value:
<pre>

    CUSTOM_ADDRESS_OVERWRITE = {
        '127.0.0.1:5000' : "omd_site_mon",
    }

</pre>

In custom server configurations use the key 'instance' to set a custom name
<pre>

    servers = [
        {
            'protocol' : 'http',
            'address'  : 'localhost',
            'port'     : 80 ,
            'instance' : "my-webserver"
        },
    ]

</pre>

Change-Id: I5b41de6f6d3b86f7b4fb3d2a8f5a734d149b597e