Remove ancient setup.sh + helpers
There was a time (before OMD was invented) where Checkmk was distributed as
standalone .tar.gz source archive. The setup.sh and autodetect.py were used
to setup Checkmk as addon for an existing Nagios/Icinga installation. Once
OMD was built, the script was called by the Checkmk package Makefile in OMD,
but most of the logic was not relevant for OMD packaging.

This commit now removes the whole unused stuff, leaving 40 Makefile lines
in place which should do the job now.

A next cleanup step to completely replace the dist archive packaging from
the top level Makefile and move the needed parts to omd/packages/check_mk/check_mk.make

Change-Id: Ibd9c402004c193fdd216501307bf2f05dd56a334