CheckAPI: add new datasource methods
The methods required in the future are similar, but have
some subtle differences regarding 'extra_sections' and
node info. In order to keep things maintainable, we
add new methods, and replace the callsites of the old
ones ony by one.

This also inprinciple implements the 'parsed_section_name' feature
described in CMK-4006 (it is not yet applied everywhere, though).

Change-Id: I5044421cbc59afc1a7166c09a3e9d53d17af2e9e