Move backup snapshot specific code to specific module
The multitar module holds two kind of independent "snapshot" logics.
The sync snapshots and the backup snapshots, that are used for the
"discard changes" feature on the "activate changes page". They do
not have a lot of shared code. For this reason we now move the
backup snapshot code to the related WATO code to have it in a single
place with the snapshot creation code.

Change-Id: I510a836919d6125b8615a5d0421792db5e7ca32c