Cleanup unit test site environment initialization
The site paths are pointing to a temporary path during unit tests. In
the last weeks a lot of tests began to initialize different paths
locally or in their own fixtures leading to a confusing situation where
tests affecting eachother made results unstable in some situations.

The needed directories and files are initialized in a central fixture in
tests/unit/conftest.py. And everything is refreshed for each test
function execution to ensure there is no test reading the files of a
previous one.

Change-Id: Ia4045276707d6e7fa972cd0d6ace7acfe55511dc