10961: Monitoring of AWS DynamoDB tables
This werk introduces monitoring of AWS DynamoDB tables. The
monitoring includes overall limits for AWS DynamoDB (both
account-wide and per-table limits on Read and Write Capacity
Units) and a summary of all tables. Furthermore, we also
monitor read and write capacities as well as latencies on a
per-table basis. To make use of this werk, you have to
configure the related AWS special agent to monitor Dynamo
DB.

Change-Id: I242504fa77d903ab7b8401f7bf658c535400ba9b