More aligning of CRE and CME: Moved common logic to base class
With the previous cleanups it is now possible to move additional common
logic to the base class of the snapshot managers. This time we have
unified the entry point, the generate_snapshots method.

This gives a clearer view on the similarities and the differences.

We now have most of the common logic in sync. Next we can change the CRE
to produce a site config sync directory, just like the CME does. Then we
have a clear 2-step snapshot creation (1: collect, 2: pack).

With that in place we will be able to a) keep the 2nd step in place just
for 1.6 compatibility and b) add the new sync implementation without
performing the 2nd step.

Change-Id: I456f540bc2d76b52104dee2b1f1a830a2fcba57f