CRE snapshot is now created in two steps
The first steps moves all needed files to a site_config directory.  This
step is mandatory for the site synchronization. This now done in the var
hierarchy. The site_config directories will be updated incrementally in
the future to optimize the synchronization.

The second step now packs the synchronization snapshot tar archive.
This step will be optional and version dependent in the future. Only
synchronizations to old remote sites (which are not aware of the new
synchronization API) will use this second step.

Change-Id: I8f6bf747b0c36dd5a3b3ba9734cf35b560ca8ac2