11089 FIX AWS ELB & ELBv2 host discovery based on user-specified names
This werk fixes a bug with the discovery of AWS ELB/ELBv2 load balancers based
on load balancer names specified by the user. In the case where both the overall
services tags and specific names (which are supposed to overrule the overall tags)
for ELB or ELBv2 were specified, all ELB/ELBv2 load balancers were fetched, ignoring
the user-specified names.

Change-Id: I0bfe7fa2744adfdcde83c809f79df521397ff362