Start cleaning up old snapshot component parsing
Since we have a better data structure, a named tuple, for the
snapshot components (aka replication paths), the snapshot component
parsing is not needed anymore. The following commits will clean that
up.

Start moving some specific logic to the specific call sites to reduce
the features of the snapshot parser. This makes it easier to clean it
up in a following step.

Change-Id: Ib418d60303fdf9215a2e83b39727b2c51519d197