Move SNMP backends out of base  #1
They are only used in `cmk.fetchers`, which should not depend on base.
We therefore move the backends under `cmk.fetchers` and the types that
they use.

There is not much risk here: pylint and mypy will identify wrong
imports.  The change is nevertheless going to be large enough that we
make it in several commits.

CMK-4218

Change-Id: I6560f7bbe9811a70787b33cde2d535efae6d8dd4