distributed_wato.mk is now created by the central site
To make the incremental configuration creation possible we need to
centralize the whole configuration file creation. The
distributed_wato.mk was created on the remote sites after applying the
configuration snapshot.

From now on this file is created by the central site and will be
replicated to the remote site like the other config files.

Change-Id: I70db78789b042ad9fe69bc187f5710d470b0739d