sitespecific.mk files are now created by the central site
To make the incremental configuration creation possible we need to
centralize the whole configuration file creation. The sitespefic.mk
files that hold the site specific global settings of the various config
domains was created on the remote sites after applying the configuration
snapshot.

From now on these files are created by the central site and will be
replicated to the remote site like the other config files. The old
behaviour is kept for compatibility with pre 1.7 remote sites.

Change-Id: I1c1aa626655eb15606948bf36acb54aa09e0f11c