New config sync: Implement automation for first sync step
Called on remote site from a central site to get the current config sync
state.

The central site hands over the list of replication paths it will try to
synchronize later.  The remote site computes the list of replication
files and sends it back together with the current configuration
generation ID. The config generation ID is increased on every WATO
modification and ensures that nothing is changed between the two config
sync steps.

Change-Id: Ia8c967009b3b906afe4bfc68ec61a5527d2a97b9