Move console logging under `cmk.util.log`
As a user, it makes sense to get console logging from `log.console`.

The functions remaining from the previous `cmk.base.console` are also
moved to dedicated modules.

This allows packages that are not in base (like the fetchers) to log to
the console using our own logging facility.

Regular log functions are still imported from `cmk.utils.log`, console
logging is importer from `cmk.utils.log.console`.

The not so useful standard stream wrapper is moved to `obsolete_output`.

The `section` FSM is moved to `cmk.base.section`.

CMK-4243

Change-Id: Id1c2f808d776efeee46a4d094994674142b5f2a9