New config sync: Add sync receiving logic
The new automation call "receive-config-sync" has been added.

The central site hands over the a tar archive with the files to be
written and a list of files to be deleted. The configuration generation
is used to validate that no modification has been made between the two
sync steps (get-config-sync-state and this autmoation).

Change-Id: I721ccbacbafd4dd2a67357ec7bff5bb61864f1a8