New config sync: Put glue between central and remote site
The full config synchronization chain is now in place. Most of the logic
is already working. Now that the >1.7 synchronization is working, the
detection of the remote sites version has been enabled again, which
means that the new logic will be used for 1.7 or newer sites and the
older logic will be used for older sites.

Next steps:

a) Add a test for the full >1.7 synchronization procedure
b) Bring the configuration generation to life

Change-Id: Ib3c1a370fed3ce3980d292a1e526302fb319dbff