New config sync: Introduce config generation
The new config sync is done in two steps:

a) ask for the remote site step
b) compute files to sync and sync

In case the remote config is modified between these steps we need
some mechanism to prevent the second step on the remote site.

The config generation is a simple number that is increased on each
configuration change that is done by WATO.

The first call gets the current configuration generation with the
first response. The second request sends the configuration generation
back and both are compared. This way we ensure that we have a
consistent state during the whole config sync procedure on the
remote site.

Change-Id: I4a2aaa050f4ede7e9cafbdb4399fb379d78af003