Make recent pylint happy with ManualCheckParameterRulespec.
The Rulespec class hierarchy is quite confusing to human readers and tools:
The constructors vary quite a lot, sometimes a parameter is a kwarg,
sometimes it's a simple arg, sometimes it's there, sometimes not. And to
make things even more interesting, the order of arguments varies, too.

OTOH, in register_rule() we pretend that we can handle those constructors in
a consistent way, which was simply not true previously.  This commit fixes
that problem by adding match_type/name/factory_default kwargs with
compatible defaults to ManualCheckParameterRulespec.__init__.

When we use Python 3, we should probably make the epic argument lists
kwargs-only, which is the only sane way to use them (and we do this already
most of the time, anyway).

Change-Id: Ibb84248ab33afcb8adce9e545c17cceef0151a1a