distributed.mk of dcd is now also created by the central site
To make the incremental configuration creation work we need to
centralize the whole configuration file creation. The distributed.mk
files that tell the DCD whether or not a connection is working in
"remote site" mode was previously created dynamically by the remote
site in the moment a config snapshot was received.

From now on these files are created by the central site and will be
replicated to the remote site like the other config files. The old
behaviour is kept for compatibility with pre 1.7 remote sites.

Change-Id: I4fe3dcbbb0aabab95a39afd6f4ae4773dc21e009