11039 Kubernetes: new option to add a namespace prefix
The special agent for Kubernetes now supports an option
to add the namespace of namespaced objects as a prefix
to the hostname in Checkmk. This makes it possible to
distinguish objects with the same name in different
namespaces. A pod inside the namespace "prod" with the
name "redis" would then e.g. be added with the hostname
"prod_pod_redis". You don't need to activate this
option if you don't use namespaces in your cluster.

Change-Id: Id234c70bb138ac2164c53b0cbb557fcd7d7adac5