11073 FIX NagVis snapin: Fix empty content during refresh
When the NagVis snapin is used with a larger number of maps or objects on these
maps, the refresh of the map states takes some time. During longer refresh
times, the snapin was not displaying the maps at all.

The already known state is now shown during refresh until the new information
is available.

Change-Id: I2c25028fad95bc6684161ebefc39afb6a3f4b56a