Remove timestamp from config file contents
Even if the timestamp in the file content is useful e.g. for trouble
shooting, it now needs to be removed from the file content (at least for
config files of the CME).

Background is that with the new incremental sync only as few files as
possible should be synchronized during the Activate change to save time
and bandwidth.

In the CME, however, many files are newly created ad hoc. Therefore,
these files always contain new timestamps, even if the content has not
changed at all. These files are therefore always synchronized.

The easiest way to improve this is to remove the time stamps from the
files. A little less information during the troubleshooting we now take
into account.

Change-Id: I96ad371463d2eae53891c4a0573b6c477c712d1a