Work around commandline length limits for mypy invocation.
Alas, there seems to be no way to tell mypy to read the files to check from
a file.  :-/  As a consequence, we have to revert to an xargs hack, making
sure that we do not overwrite an output log with failures.  Downside: We
might not get all errors in a single run.

This is all quite horrible, and in the long run, we should somehow remove
the find-python-files technology, making the checking more modular.

Change-Id: I4a9fb2374b24ea5ced3e3dd191cad3bfac4abed4