11044 FIX dell_idrac_fans: don't crash on unknown device states
The check plugin crashed when fans of a device were in
states that did not return a device reading. The plugin
now returns only the status and the name for the states
"other", "unknown", and "failed".

Fans that are in the states "unknown" or "other" will
now no longer be discovered. Fans that are in the state
"failed" will still be discovered.

Change-Id: Ic8640d5512243380acb4d25d7eba0a93d58d9280