discovery: explicitly pass plugin whitelist along I
In '_discover_services', we stop calling sources.get_check_plugin_names.
Instead we pass the the whitelist explicitly along, and only look at check
plugin candidates that the host sections have at least *some* data for.

The clean up of the now unneeded arguments is done in the next commit.

Change-Id: Ic36a51f4ca9033e4315dcb3fff580b9548132444