11150 Refactoring of the monitoring of Cisco VPN sessions
This werk refactors the code for monitoring the number of Cisco
VPN sessions. The three checks cisco_asa_ipsecsessions,
cisco_asa_svcsessions and cisco_asa_webvpnsessions are replaced
with one check with three items (cisco_vpn_sessions). The
corresponding rule for defining upper levels on the number of active
sessions is renamed from "Cisco SVC/WebVPN/IPsec Sessions" to "Cisco
VPN Sessions"

This change is incompatible. Users have to re-discover the services
of affected hosts and re-create the rules for the upper levels on
the number of active sessions. Furthermore, any previously created rules
should be deleted, since they are deprecated as of this werk and have
no more effect.

Change-Id: I4728c01cbedd7e02068501e4bf15e4b286d5d8d5