Start update source container without tmpfs
For some reason the source container is not able to start anymore on our
dc14 build system while it is still working on dc44. Was not able to
find the reason for this issue.
However, it is possible to workaround this issue by not using a tmpfs
for the sites tmp directory.  Since the tmpfs is an optional performance
feature, this workaround seems to be OK.

Change-Id: I143e0fb298052498db0f0126ef1f8b14879d343c