Move related types close to each other
OIDCached and OIDBytes _are_ OIDSpec, defining them under a completely
different hierarchy is confusing.

The obsolete `OID_*` enum should be superseded by OIDSpec.  It is only
logical to have them close to each other.

Finally, these are types that are used both in the fetchers (via
`snmp.get_snmp_table()` that is nor to be moved) and base.

CMK-4504

Change-Id: I956771495b549e6271db28e8f0b574016e977778