Remove WSGI sub interpreters per language
Before 1.6, Checkmk had many localizable texts that were resolved in the
module namespace, i.e. at the time a python module was imported. This
made it necessary to have a separate Python interpreter for each
language so that different users could use the GUI in different
languages.

Now all localizable texts are evaluated during the processing of a
request, not when the application is loaded. If this does not apply to
some texts, this is to be seen as a bug and must be fixed.

Because the localized texts are now evaluated during the request, we can
remove the sub-interpreters per language. This saves memory and
simplifies the setup a bit.

Change-Id: I3bfba1aa363dd5e766a8b31a993b6a351663dafc