Move ABCSNMPBackend to `type_defs`
We should be able to use this in `cmk.utils.snmp_table` in order to
break the last (I believe) layering violation introduced by this series
of patches.

Note that importing `fetchers` in `base` is not a layering violation.
The other way around is one.  In the end, we should have something like:

   +---------------------+
   |       base          |
   +--------+------------+
   |        |   fetchers |
   |        +------------+
   |   utils             |
   +---------------------+

which is probably ok.

CMK-4504

Change-Id: Id78e1d57181ce4aefeb7e59a238200b791c1dac5