10847 FIX mk_oracle: Broken instance specific login in some cases
In some cases the SID as it is displayed on the host process list and the
SID specific login data does not match if the matching is case sensitive.
As a consequense the variable for the specific login has not been found,
e.g. if the process is lower case but the config uses upper case SID.

Change-Id: I1fbee03e8103a67765c6ba2994d713c06407f7ef