Fix calculation if issued rds_licenses
RDS license servers are able to issue more licenses than added to their database.
The calculation 'pack_total - pack_avail' doesn't reflect, if more licenses are consumed, than available.
For the instance, this would work with the current check:
- 600 licenses in total
- 450 consumed
- 150 available
Using the existing calculation, it would result in 600-150=450 used licenses. 

But here comes the tricky part:
- 800 licenses in total
- 801 consumed
- 0 available
Using the existing calculation, it would result in 800-0=0 
With the proposed solution, it would properly reflect the issued licenses, which are handed over via agent plugin anyhow.