Drop plugin reloading after language change
Since 1.6 we require gettext localized strings to be moved away from
module import scope. Strings that need to be declared in this situation
need to use speaklater wrapped strings (`cmk.gui._l()`).

This change now finally stops supporting old localized strings declared
at import time. It will break nothing, but might make strings
non-localizable which were before.

To find these strings we will fail hard in the future in case _() is
called during import. But for the moment we keep it like this to remain
focused on cleaning up the plugin import logic.

Change-Id: Ia03cd43a9a78a045453ac8b2a12e645246d60bc3