11146 mkp_prometheus: addition of mkp version 1.1
This change updates the optional extension package for the Prometheus suite. It introduces
the multi-nodes processing for the Node Exporter. This allows Checkmk to handle a Prometheus
instance where multiple Node Exporters are connected. This mkp can be installed
directly through site. It is listed as an optional extension package in the Extension Packages
page.

Change-Id: Icd37d5d15ce0c6353e1ac98a55f72513e2990808