Fixes for mypy 0.780.
Note that some fixes are actually bugs, and we should have a look if the
same bugs are present in the 1.6.0 branch, too.

Another note: Adding asserts all over the place doesn't look right, we
should better improve our typing.  But this would involve actually
parsing/validating the values we get from exec/literal_eval/Livestatus.
Even if don't do this in its full glory, we should at least have a thin
typed layer above the untyped low-level-pray-and-hope-for-the-best code.

Change-Id: Ie65fc8b5cbec61e242c8205a9458ef61028a0642