Move `cleanup` from base to utils
The module itself has no deps on base.

With this change, we get rid of all the dependencies of `cmk.base.snmp`
on `base`. Now, we are able to increase the cohesion of our SNMP code by
moving `cmk.utils.snmp_table` and `cmk.base.snmp` back to a common
package, possibly outside of `cmk.utils`.

CMK-4610

Change-Id: I18654d4883c9c4e19671dec7e4fe4aacad795876