Improve plugin-api reference creation
Instead of the previous approach to explicitly scan the individual plugin
submodule namespace we now execute sphinx-apidoc on the whole cmk module
hierarchy and filter out the unwanted files later.

The advantage is, that it will automatically add new chapters for new
plugin APIs. It also gives sphinx an idea about the real full module name

Change-Id: Ida31cf649248ba7a1c6097a9ec398b6880aa6702