Try to fix SNMP integration tests
The backend and snmp_config fixture were both parameterized with the
backend to be used, but the patching of  cmk.utils.paths.snmpwalks_dir
was only triggered within the snmp_config fixture.

In addition, the parameters of the two fixtures were multiplied out,
which is probably not the intention.

snmp_config no longer has its own parameters and uses the instances of
the backend fixture.

Change-Id: I28085235b7a3cbe086a406925ccec13b1227cff4