Refactoring of popup_menu.js #2
This commit reduces the amount of arguments that
are passed to the function toggle_popup. Since
many of the arguments are mutual exclusive they
are now bundled together in render_popup_trigger
and submitted together as one object to
toggle_popup. The type property is added to the
object to make the code in toggle_popup more
readable.

The type "colorpicker" needed a little more care
since it included a "hidden" function call in
valuespec.py. This call is now moved to directly
to toggle_popup. This made it possible to reduce
the arguments of generate_colorpicker_body.

Change-Id: I28a3e48d5c2720205c6cce7ca1e1baef08be560f