11725 FIX storcli_pdisks, storcli_vdisks: Support multiple RAID controllers per system
When monitoring LSI MegaRAID controller via StorCLI, now all drives
of all RAID controllers of a system do get discovered and added as a service.
Previously, the discovery did only find the first controller of a system.

To differentiate between the controllers, the item name is prepended by "Cn.",
where n is the number of the controller.<br>
E.g.:
<ul>
<li><tt>RAID Virtual Drive 0/0</tt> would become <tt>RAID Virtual Drive C0.0/0</tt>
for controller 0</li>
<li><tt>RAID PDisk EID:Slot-Device 13:9-19</tt> would become
<tt>RAID PDisk EID:Slot-Device C1.13:9-19</tt> for controller 1</tt></li>
</ul>
As this Werk changes the item name, this is an incompatible change and will
require re-discovering the services.

PR from CMK-4693

Change-Id: I6a6d0cc6889d3aac35355d5a44edecf3e8e5085c